maybe this solution is even cleaner
