+++ client/allocrunner/taskrunner/template/template.go
@@ -18,6 +18,7 @@ import (
 
 	ctconf "github.com/hashicorp/consul-template/config"
 	"github.com/hashicorp/consul-template/manager"
+	"github.com/hashicorp/consul-template/signals"
 	envparse "github.com/hashicorp/go-envparse"
 	"github.com/hashicorp/go-hclog"
 	multierror "github.com/hashicorp/go-multierror"
@@ -158,6 +159,65 @@ func (c *TaskTemplateManagerConfig) Validate() error {
 	}
 
 	return nil
+}
+
+func NewTaskTemplateManager(config *TaskTemplateManagerConfig) (*TaskTemplateManager, error) {
+	// Check pre-conditions
+	if err := config.Validate(); err != nil {
+		return nil, err
+	}
+
+	tm := &TaskTemplateManager{
+		config:     config,
+		shutdownCh: make(chan struct{}),
+	}
+
+	// Parse the signals that we need
+	for _, tmpl := range config.Templates {
+		if tmpl.ChangeSignal == "" {
+			continue
+		}
+
+		sig, err := signals.Parse(tmpl.ChangeSignal)
+		if err != nil {
+			return nil, fmt.Errorf("Failed to parse signal %q", tmpl.ChangeSignal)
+		}
+
+		if tm.signals == nil {
+			tm.signals = make(map[string]os.Signal)
+		}
+
+		tm.signals[tmpl.ChangeSignal] = sig
+	}
+
+	// Build the consul-template runner
+	runner, lookup, err := templateRunner(config)
+	if err != nil {
+		return nil, err
+	}
+	tm.runner = runner
+	tm.lookup = lookup
+
+	go tm.run()
+	return tm, nil
+}
+
+// Stop is used to stop the consul-template runner
+func (tm *TaskTemplateManager) Stop() {
+	tm.shutdownLock.Lock()
+	defer tm.shutdownLock.Unlock()
+
+	if tm.shutdown {
+		return
+	}
+
+	close(tm.shutdownCh)
+	tm.shutdown = true
+
+	// Stop the consul-template runner
+	if tm.runner != nil {
+		tm.runner.Stop()
+	}
 }
 
 // SetDriverHandle sets the executor
@@ -920,6 +980,25 @@ func newRunnerConfig(config *TaskTemplateManagerConfig,
 	conf.RendererFunc = RenderFn(config.TaskID, sandboxDir, sandboxEnabled)
 	conf.Finalize()
 	return conf, nil
+}
+
+func isSandboxEnabled(cfg *TaskTemplateManagerConfig) bool {
+	if cfg.ClientConfig != nil && cfg.ClientConfig.TemplateConfig != nil && cfg.ClientConfig.TemplateConfig.DisableSandbox {
+		return false
+	}
+	return true
+}
+
+type sandboxConfig struct {
+	thisBin     string
+	sandboxPath string
+	destPath    string
+	sourcePath  string
+	perms       string
+	user        string
+	group       string
+	taskID      string
+	contents    []byte
 }
 
 // loadTemplateEnv loads task environment variables from all templates.
